requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.1.8
cryptography==46.0.6
exceptiongroup==1.3.1
fastapi==0.128.8
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.1
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
googleapis-common-protos==1.74.0
gspread==6.2.1
h11==0.16.0
httplib2==0.31.2
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.0.2
oauthlib==3.3.1
pandas==2.3.3
pillow==11.3.0
proto-plus==1.27.2
protobuf==6.33.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
requests==2.32.5
requests-oauthlib==2.0.0
six==1.17.0
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.39.0
